+++ packages/create-package/template/bundle.esm.js
@@ -12,7 +12,7 @@ import { getFirstDayOfWeek } from "@ui5/webcomponents-base/dist/config/FormatSet
 import "./dist/Assets.js";
 
 // Import your web components here from the dist/ directory
-import "./dist/INIT_PACKAGE_VAR_CLASS_NAME.js";
+import "./dist/{{INIT_PACKAGE_VAR_CLASS_NAME}}.js";
 
 window["sap-ui-webcomponents-bundle"] = {
 	renderFinished,