+++ packages/main/cypress/specs/Form.cy.tsx
@@ -4,6 +4,7 @@ import FormItem from "../../src/FormItem.js";
 import FormGroup from "../../src/FormGroup.js";
 import Label from "../../src/Label.js";
 import Text from "../../src/Text.js";
+import Title from "../../src/Title.js";
 import Input from "../../src/Input.js";
 
 describe("General API", () => {
@@ -196,6 +197,27 @@ describe("General API", () => {
 			.and("have.prop", "emptySpanM", 0)
 			.and("have.prop", "emptySpanL", 1)
 			.and("have.prop", "emptySpanXl", 1);
+	});
+
+	it("tests calculated state of Form item-spacing='Large'", () => {
+		cy.mount(<Form id="addressForm" item-spacing="Large">
+			<FormGroup id="formGroup">
+				<FormItem id="formItem">
+					<Label slot="labelContent">Name:</Label>
+					<Text>Red Point Stores</Text>
+				</FormItem>
+
+				<FormItem>
+					<Label slot="labelContent">Country:</Label>
+					<Text>Germany</Text>
+				</FormItem>
+			</FormGroup>
+		</Form>);
+
+		cy.get("#formGroup")
+			.should("have.prop", "itemSpacing", "Large");
+		cy.get("#formItem")
+			.should("have.prop", "itemSpacing", "Large");
 	});
 
 	it("tests calculated state of two FormGroups in layout='S1 M2 L3 XL4'", () => {
@@ -786,21 +808,21 @@ describe("Accessibility", () => {
 
 						<FormItem>
 							<Label id="cityLbl" for="cityInp" slot="labelContent">ZIP Code/City:</Label>
-							<Input id="cityInp" value="411" accessible-name-ref="cityLbl"></Input>
-							<Input value="Maintown" accessible-name-ref="cityLbl"></Input>
+							<Input id="cityInp" value="411" accessibleNameRef="cityLbl"></Input>
+							<Input value="Maintown" accessibleNameRef="cityLbl"></Input>
 						</FormItem>
 					</FormGroup>
 
 					<FormGroup headerText="Contact">
 						<FormItem>
 							<Label id="streetLbl" for="streetInp" slot="labelContent">Street:</Label>
-							<Input id="streetInp" value="Main St" accessible-name-ref="streetLbl"></Input>
-							<Input id="streetNumberInp" value="1618" accessible-name-ref="streetLbl"></Input>
+							<Input id="streetInp" value="Main St" accessibleNameRef="streetLbl"></Input>
+							<Input id="streetNumberInp" value="1618" accessibleNameRef="streetLbl"></Input>
 						</FormItem>
 
 						<FormItem>
 							<Label id="countryLbl" for="countrySel" slot="labelContent">Country:</Label>
-							<Input id="countrySel" accessible-name-ref="countryLbl"></Input>
+							<Input id="countrySel" accessibleNameRef="countryLbl"></Input>
 						</FormItem>
 					</FormGroup>
 				</Form>
@@ -813,19 +835,19 @@ describe("Accessibility", () => {
 
 					<FormItem>
 						<Label id="cityLbl2" for="cityInp2" slot="labelContent">ZIP Code/City:</Label>
-						<Input id="cityInp2" value="411" accessible-name-ref="cityLbl2"></Input>
-						<Input value="Maintown" accessible-name-ref="cityLbl2"></Input>
+						<Input id="cityInp2" value="411" accessibleNameRef="cityLbl2"></Input>
+						<Input value="Maintown" accessibleNameRef="cityLbl2"></Input>
 					</FormItem>
 
 					<FormItem>
 						<Label id="streetLbl2" for="streetInp2" slot="labelContent">Street:</Label>
-						<Input id="streetInp2" value="Main St" accessible-name-ref="streetLbl2"></Input>
-						<Input id="streetNumberInp" value="1618" accessible-name-ref="streetLbl2"></Input>
+						<Input id="streetInp2" value="Main St" accessibleNameRef="streetLbl2"></Input>
+						<Input id="streetNumberInp" value="1618" accessibleNameRef="streetLbl2"></Input>
 					</FormItem>
 
 					<FormItem>
 						<Label id="countryLbl2" for="countrySel2" slot="labelContent">Country:</Label>
-						<Input id="countrySel2" accessible-name-ref="countryLbl2"></Input>
+						<Input id="countrySel2" accessibleNameRef="countryLbl2"></Input>
 					</FormItem>
 				</Form>
 			</>);
@@ -855,4 +877,66 @@ describe("Accessibility", () => {
 		cy.get("#nameInp")
 			.should("be.focused");
 	});
+});
+
+
+ui5AccDescribe("Automated accessibility tests", () => {
+	it("with header text", () => {
+		cy.mount(
+			<Form headerText="Address">
+				<FormItem>
+					<Label slot="labelContent">Name:</Label>
+					<Text>Red Point Stores</Text>
+				</FormItem>
+				
+				<FormItem>
+					<Label slot="labelContent">ZIP Code/City:</Label>
+					<Text>411 Maintown</Text>
+				</FormItem>
+				
+				<FormItem>
+					<Label slot="labelContent">Street:</Label>
+					<Text>Main St 1618</Text>
+				</FormItem>
+
+				<FormItem>
+					<Label slot="labelContent">Country:</Label>
+					<Text>Germany</Text>
+				</FormItem>
+			</Form>
+		);
+
+		cy.ui5CheckA11y();
+	})
+
+	it("with custom header", () => {
+		cy.mount(
+			<Form>
+				<div slot="header">
+					<Title>Address</Title>
+				</div>
+				<FormItem>
+					<Label slot="labelContent">Name:</Label>
+					<Text>Red Point Stores</Text>
+				</FormItem>
+				
+				<FormItem>
+					<Label slot="labelContent">ZIP Code/City:</Label>
+					<Text>411 Maintown</Text>
+				</FormItem>
+				
+				<FormItem>
+					<Label slot="labelContent">Street:</Label>
+					<Text>Main St 1618</Text>
+				</FormItem>
+
+				<FormItem>
+					<Label slot="labelContent">Country:</Label>
+					<Text>Germany</Text>
+				</FormItem>
+			</Form>
+		);
+
+		cy.ui5CheckA11y();
+	})
 });